author relationship established

+++ BookStoreApi/BookStoreApi/Application/AuthorOperations/Commands/DeleteAuthor/DeleteAuthorCommand.cs
@@ -1,4 +1,6 @@
-﻿using BookStoreApi.DBOperations;
+﻿using System;
+using System.Linq;
+using BookStoreApi.DBOperations;
 
 namespace BookStoreApi.Application.AuthorOperations.Commands.DeleteAuthor
 {
@@ -14,10 +16,13 @@ namespace BookStoreApi.Application.AuthorOperations.Commands.DeleteAuthor
 
         public void Handle()
         {
-            var author = _dbContext.Authors.SingleOrDefault(x=> x.Id == AuthorId);
+            var author = _dbContext.Authors.SingleOrDefault(x => x.Id == AuthorId);
             if (author is null)
                 throw new InvalidOperationException("Yazar Bulunamadı");
-            
+
+            if (author.Books.Any())
+                throw new InvalidOperationException("Yazarın yayınlanmış kitapları bulunuyor. Önce kitapları silin.");
+
             _dbContext.Authors.Remove(author);
             _dbContext.SaveChanges();
         }
